chore(product-analytics): move temporal code into the product
The upgrade-queries backfill workflow, its activities and its tests now live
under products/product_analytics/backend/temporal/, next to the Insight model
they rewrite. Over 20 products already keep a backend/temporal/.

Core reaches the workflow through a new backend/facade/temporal.py, the same
channel error_tracking and data_quality use, so the product keeps its
backend:contract-check narrowing.

backend/temporal/** joins the contract-check inputs because the isolation lint
watches that wiring location by presence. The CD deploy filter for the general
purpose temporal worker follows the code to its new path.

Refs #77396
