+++ .github/workflows/container-images-cd.yml
@@ -475,7 +475,7 @@ jobs:
               id: check_changes_general_purpose_temporal_worker
               run: |
                   # Changes to this detector must trigger once so newly covered code ships with the worker.
-                  if git diff --name-only HEAD^ HEAD | grep -qE '^products/notebooks/backend/|^products/exports/backend/temporal/subscriptions|^products/engineering_analytics/backend/|^products/canvas/backend/|^posthog/llm/|^posthog/temporal/common|^posthog/temporal/dlq_replay|^posthog/temporal/session_recordings|^posthog/temporal/proxy_service|^posthog/temporal/(ai_observability|llm_analytics)|^posthog/temporal/product_analytics|^posthog/temporal/delete_persons|^posthog/temporal/delete_teams|^posthog/temporal/ingestion_acceptance_test|^posthog/temporal/backfill_group_type_created_at|^posthog/management/commands/start_temporal_worker.py$|^posthog/kafka_client/|^pyproject.toml$|^bin/temporal-django-worker$|^posthog/clickhouse/migrations/' \
+                  if git diff --name-only HEAD^ HEAD | grep -qE '^products/notebooks/backend/|^products/exports/backend/temporal/subscriptions|^products/engineering_analytics/backend/|^products/canvas/backend/|^posthog/llm/|^posthog/temporal/common|^posthog/temporal/dlq_replay|^posthog/temporal/session_recordings|^posthog/temporal/proxy_service|^posthog/temporal/(ai_observability|llm_analytics)|^products/product_analytics/backend/temporal/|^posthog/temporal/delete_persons|^posthog/temporal/delete_teams|^posthog/temporal/ingestion_acceptance_test|^posthog/temporal/backfill_group_type_created_at|^posthog/management/commands/start_temporal_worker.py$|^posthog/kafka_client/|^pyproject.toml$|^bin/temporal-django-worker$|^posthog/clickhouse/migrations/' \
                       || git diff HEAD^ HEAD -- .github/workflows/container-images-cd.yml | grep -qE '^\+.*products/exports/backend/temporal/subscriptions'; then
                       echo "changed=true" >> $GITHUB_OUTPUT
                   else
