+++ posthog/test/repo_invariants/dataclass_frozen_baseline.txt
@@ -76,7 +76,6 @@
 1 posthog/temporal/ingestion_acceptance_test/test_cases_discovery.py
 1 posthog/temporal/mcp_analytics/intent_clustering/coordinator.py
 1 posthog/temporal/mcp_analytics/intent_clustering/team_discovery.py
-1 posthog/temporal/product_analytics/upgrade_queries_workflow.py
 1 posthog/temporal/session_replay/rasterize_recording/activities/stuck_counter.py
 1 posthog/temporal/tests/common/test_logger.py
 1 posthog/temporal/tests/data_modeling/test_log_routing.py
@@ -163,6 +162,7 @@
 1 products/posthog_ai/eval_harness/scorers/tracing.py
 1 products/product_analytics/backend/hogql_queries/funnels/funnel_event_query.py
 1 products/product_analytics/backend/hogql_queries/paths/test/test_paths_query_runner.py
+1 products/product_analytics/backend/temporal/upgrade_queries_workflow.py
 1 products/product_tours/backend/generate_tour_content/generate.py
 1 products/pulse/backend/temporal/activities.py
 1 products/replay_vision/backend/tests/test_gemini_cleanup_activities.py
