+++ products/model_crossing_uses_baseline.txt
@@ -634,7 +634,6 @@ product_analytics.Insight posthog.models.file_system.unfiled_file_saver in-colle
 product_analytics.Insight posthog.models.product_intent.product_intent instance-many(filter) 2
 product_analytics.Insight posthog.models.resource_transfer.visitors.insight other(Return) 1
 product_analytics.Insight posthog.models.sharing_configuration passed-as-class(cast) 1
-product_analytics.Insight posthog.temporal.product_analytics.upgrade_queries_activities instance-many(filter) 1
 product_analytics.Insight products.actions.backend.api.action instance-many(extra) 1
 product_analytics.Insight products.alerts.backend.max_tools instance-many(filter) 1
 product_analytics.Insight products.alerts.backend.max_tools passed-as-class(cast) 1
