+++ products/product_analytics/backend/temporal/tests/conftest.py
@@ -1,0 +1,9 @@
+import pytest
+
+from temporalio.testing import ActivityEnvironment
+
+
+@pytest.fixture
+def activity_environment():
+    """Return a testing temporal ActivityEnvironment."""
+    return ActivityEnvironment()
