+++ products/product_analytics/turbo.json
@@ -8,6 +8,8 @@
                 "backend/routes.py",
                 "backend/models/**",
                 "backend/migrations/**",
+                "backend/temporal/**",
+                "!backend/temporal/tests/**",
                 "backend/hogql_queries/trends/**",
                 "!backend/hogql_queries/trends/test/**"
             ],
